Conflicts:
	features/com.example.addressbook.feature/feature.xml
	plugins/com.example.addressbook.editing/META-INF/MANIFEST.MF
	plugins/com.example.addressbook/META-INF/MANIFEST.MF
	plugins/com.example.addressbook/src/com/example/addressbook/internal/address/list/AddressDeleteHandler.java
	plugins/com.example.addressbook/src/com/example/addressbook/internal/address/list/AddressListViewPart.java

+++ plugins/com.example.addressbook/src/com/example/addressbook/internal/address/list/AddressDeleteHandler.java
@@ -1,14 +1,16 @@
 package com.example.addressbook.internal.address.list;
 
+import java.util.Iterator;
+
 import org.eclipse.core.commands.AbstractHandler;
 import org.eclipse.core.commands.ExecutionEvent;
 import org.eclipse.core.commands.ExecutionException;
 import org.eclipse.jface.viewers.ISelection;
+import org.eclipse.jface.viewers.IStructuredSelection;
 import org.eclipse.ui.handlers.HandlerUtil;
 
 import com.example.addressbook.entities.Address;
 import com.example.addressbook.services.IAddressService;
-import com.example.addressbook.utils.SelectionUtils;
 
 /**
  * Handler to delete selected addresses.
@@ -19,9 +21,17 @@ public class AddressDeleteHandler extends AbstractHandler {
 
 	public Object execute(ExecutionEvent event) throws ExecutionException {
 		ISelection selection = HandlerUtil.getCurrentSelection(event);
-		for (Address address : SelectionUtils.getIterable(selection, Address.class)) {
-			addressService.deleteAddress(address.getId());
+
+		if (selection instanceof IStructuredSelection) {
+			for (Iterator<?> i = ((IStructuredSelection) selection).iterator(); i.hasNext();) {
+				Object selectedObj = i.next();
+				if (selectedObj instanceof Address) {
+					Address address = (Address) selectedObj;
+					addressService.deleteAddress(address.getId());
+				}
+			}
 		}
+
 		return null;
 	}
